+++ react/components/Statement/Atoms/SubjectAtom.js
