+++ react/components/Statement/Atoms/VerbAtom.js
