+++ react/components/Statement/index.js
